refactor(tui): split calendar_dialog.go below 1000 lines
Move constructor, update, and view into sibling files in package tui.
No behavior change.

Related: #652
