+++ internal/tui/calendar_dialog.go
@@ -6,9 +6,7 @@ import (
 	"time"
 
 	"charm.land/bubbles/v2/help"
-	"charm.land/bubbles/v2/key"
 	"charm.land/bubbles/v2/viewport"
-	tea "charm.land/bubbletea/v2"
 	lipgloss "charm.land/lipgloss/v2"
 
 	"github.com/douglasdemoura/chroncal/internal/account"
@@ -236,419 +234,6 @@ type CalendarDialogModel struct {
 	// calendars, nil for local calendars and create mode. Enter on it emits
 	// AccountSettingsRequestedMsg so the owning manager/host can drill in.
 	accountOpener *OpenerField
-}
-
-// NewCalendarDialogModel builds a dialog for create (params.ID==0) or edit.
-func NewCalendarDialogModel(params CalendarDialogParams, theme Theme) CalendarDialogModel {
-	title := "New local calendar"
-	if params.ID > 0 {
-		title = "Edit calendar"
-		// Apple's "Get Info" sheet shows the default badge inline with the
-		// title — readable at a glance and impossible to confuse with an
-		// editable field. We do the same so users opening a calendar via
-		// the sidebar's Return keypress see the state immediately.
-		if params.IsDefault {
-			title += " · Default"
-		}
-	}
-	if params.Color == "" {
-		params.Color = "#a6e3a1"
-	}
-
-	styles := DefaultDialogStyles()
-	dialog := NewDialog(title, styles)
-	dialog.SetWidth(62)
-
-	formStyles := DefaultFormStyles()
-	formStyles.LabelLayout = LabelInline
-	formStyles.ShowFocusMarker = true
-	formStyles.ButtonAlign = ButtonAlignRight
-	formStyles.ButtonRule = true
-
-	nameField := NewTextField("e.g. Work")
-	nameField.SetValue(params.Name)
-	nameField.SetCharLimit(256)
-
-	colorField := NewColorField(theme.CalendarSwatches, params.Color, theme.TextDim)
-
-	descField := NewTextField("Shared family schedule")
-	descField.SetValue(params.Description)
-	descField.SetCharLimit(500)
-
-	emailField := NewTextField("you@example.com")
-	emailField.SetValue(params.OwnerEmail)
-	emailField.SetCharLimit(256)
-
-	items := []FormItem{
-		{Label: "Name", Field: nameField, Required: true},
-		{Label: "Color", Field: colorField, Required: true},
-		{Label: "Description", Field: descField},
-		{Label: "Owner email", Field: emailField},
-		{Label: "", Field: NewStaticField("", nil)},
-	}
-
-	// The dialog width isn't known until SetSize. Truncate compact account
-	// context at render time so long names and errors stay on one row.
-	contentWidth := new(int)
-
-	// staticLine builds a one-row static field: truncate to the content
-	// width first, then style — slicing after styling would cut through
-	// ANSI escapes.
-	staticLine := func(text string, style lipgloss.Style) FormItem {
-		return FormItem{Label: "", Field: NewStaticField(text, func(s string) string {
-			if *contentWidth > 0 {
-				s = truncateTo(s, *contentWidth)
-			}
-			return style.Render(s)
-		})}
-	}
-	// Edit mode surfaces the calendar's visibility and owner context as
-	// actionable rows. That is a Display Calendar checkbox (visibility is
-	// immediate and never auto-saved with metadata). And either
-	// "Location: Local" or an "Account: <name> ›" opener that drills into
-	// the owner account. Create mode has no immutable ID, so neither row
-	// applies there.
-	var (
-		visibilityCb  *CheckboxField
-		accountOpener *OpenerField
-		openerIdx     = -1
-	)
-	if params.ID > 0 {
-		visibility := NewCheckboxField("", !params.Hidden)
-		visibility.SetContent("Display calendar")
-		visibilityCb = visibility
-		items = append(items, FormItem{Label: "", Field: visibility, AlignToFieldColumn: true})
-
-		// Account and Location sit in the shared label column like every
-		// other row (Apple Settings detail-row layout); the opener's value
-		// carries the drill-in chevron.
-		if params.RemoteLinked {
-			accountName := strings.TrimSpace(params.AccountName)
-			if accountName == "" {
-				accountName = "Connected account"
-			}
-			openerIdx = len(items)
-			opener := NewOpenerField(accountName + " ›")
-			accountOpener = opener
-			items = append(items, FormItem{Label: "Account", Field: opener})
-			for _, line := range syncHealthDialogLines(params, theme) {
-				items = append(items, staticLine(line.text, line.style))
-			}
-			// Account calendars carry no Delete button. A delete here would
-			// only remove the local copy, not the account's calendar. This
-			// footnote explains why and points at the local alternative.
-			note := lipgloss.NewStyle().Foreground(theme.TextDim)
-			ownership := "This calendar lives in your " + accountName + " account."
-			if strings.TrimSpace(params.AccountName) == "" {
-				ownership = "This calendar lives in your connected account."
-			}
-			items = append(items,
-				FormItem{Label: "", Field: NewStaticField("", nil)},
-				staticLine(ownership, note),
-				staticLine("Turn off Display calendar to hide it on this device.", note),
-				staticLine("To remove it, open Account › Manage Calendars.", note),
-			)
-		} else {
-			items = append(items, FormItem{Label: "Location", Field: NewStaticField("Local", nil)})
-		}
-	}
-
-	form := NewForm("Save", formStyles, items...)
-
-	savedID := params.ID
-	saveMakeDefault := new(bool)
-	form.OnSubmit(func(f *Form) tea.Cmd {
-		nameVal := strings.TrimSpace(f.Field(cdIdxName).(*TextField).Value())
-		hexVal := strings.TrimSpace(f.Field(cdIdxColor).(*ColorField).Value())
-		descVal := strings.TrimSpace(f.Field(cdIdxDescription).(*TextField).Value())
-		emailVal := strings.TrimSpace(f.Field(cdIdxEmail).(*TextField).Value())
-
-		msg := CalendarSavedMsg{
-			ID:          savedID,
-			Name:        nameVal,
-			Color:       hexVal,
-			Description: descVal,
-			OwnerEmail:  emailVal,
-			MakeDefault: *saveMakeDefault,
-		}
-		*saveMakeDefault = false
-
-		return func() tea.Msg { return msg }
-	})
-
-	form.OnCancel(func(f *Form) tea.Cmd {
-		return func() tea.Msg { return CalendarDialogClosedMsg{} }
-	})
-
-	localDraft := params
-
-	// form is intentionally NOT set here: the button/handler wiring below
-	// mutates the local form, and m.form captures the final state just
-	// before return.
-	m := CalendarDialogModel{
-		id:                params.ID,
-		name:              params.Name,
-		linked:            params.RemoteLinked,
-		dialog:            dialog,
-		body:              viewport.New(),
-		help:              newThemedHelp(theme),
-		theme:             theme,
-		accentColor:       theme.Selected,
-		mutedColor:        theme.Muted,
-		textDimColor:      theme.TextDim,
-		saveMakeDefault:   saveMakeDefault,
-		accountConnection: false,
-		localDraft:        &localDraft,
-		contentWidth:      contentWidth,
-		hidden:            params.Hidden,
-		visibilityCb:      visibilityCb,
-		accountOpener:     accountOpener,
-	}
-	m.body.MouseWheelEnabled = true
-
-	// Edit mode, not yet default: surface "Set as Default" without forcing a
-	// trip through the manage-calendars list.
-	if params.ID > 0 && !params.IsDefault {
-		id := params.ID
-		name := params.Name
-		form.SetUtilityActionButton("Set as Default", Button, func() tea.Msg {
-			return CalendarSetDefaultRequestedMsg{ID: id, Name: name}
-		})
-	}
-
-	// Remote calendars drill into their owner account via the inline
-	// "Account: <name> ›" opener, not a separate button. The opener's
-	// Enter then emits the canonical AccountSettingsRequestedMsg the
-	// host already routes. The calendar manager intercepts that to push
-	// the account detail. It does not disturb the in-progress calendar
-	// draft.
-	if openerIdx >= 0 {
-		accountID := params.AccountID
-		capturedIdx := openerIdx
-		form.OnFieldEnter(func(f *Form, field int) tea.Cmd {
-			if field != capturedIdx {
-				return nil
-			}
-			return func() tea.Msg { return AccountSettingsRequestedMsg{AccountID: accountID} }
-		})
-	}
-
-	// Edit mode exposes Delete as a leading action that targets the
-	// calendar's immutable ID. It is destructive (ButtonDanger) and only
-	// requests removal. The host owns the safe-confirm flow. Account
-	// calendars get no Delete. The button could only drop the local copy
-	// while the account still owns the calendar. The form explains that
-	// in a footnote instead (hide locally via Display calendar; manage
-	// membership in Account ▸ Manage Calendars). Export is a manager-only
-	// affordance (the legacy app has no export handler yet). It is gated
-	// on ManagerEmbedded to avoid a no-op button in legacy-wired dialogs.
-	// Apple sheet disposition: Set as Default and Export live on the quiet
-	// utility tier above the commit row. Delete — destructive — sits in
-	// the commit row's bottom-left corner, as far as possible from Save.
-	if params.ID > 0 {
-		id := params.ID
-		name := params.Name
-		if params.ManagerEmbedded {
-			form.SetUtilityActionButton("Export Calendar…", Button, func() tea.Msg {
-				return CalendarExportRequestedMsg{ID: id, Name: name}
-			})
-			if !params.RemoteLinked {
-				form.SetUtilityActionButton("Move to Account…", Button, func() tea.Msg {
-					return CalendarMoveToAccountRequestedMsg{ID: id, Name: name}
-				})
-			}
-		}
-		if params.RemoteLinked {
-			// The keep-local counterpart to Delete: unlink from the account
-			// but keep every downloaded event as a local calendar.
-			form.SetUtilityActionButton("Keep as Local Calendar…", Button, func() tea.Msg {
-				return CalendarKeepLocalRequestedMsg{ID: id, Name: name}
-			})
-		} else {
-			form.SetLeadingActionButton("Delete Calendar…", ButtonDanger, func() tea.Msg {
-				return CalendarDeleteRequestedMsg{ID: id, Name: name}
-			})
-		}
-	}
-
-	// Create mode with at least one calendar already on disk: offer to
-	// promote the new row to default in one save. Do not force a
-	// follow-up trip through the list dialog. Suppressed for the very
-	// first calendar. The service auto-promotes that row in silence.
-	if params.ID == 0 && params.OfferDefault {
-		form.SetLeadingActionButton("Save and Set as Default", Button, func() tea.Msg {
-			return calendarSavePromotePressedMsg{}
-		})
-	}
-
-	m.form = form
-	return m
-}
-
-// NewAccountDialogModel opens account sign-in directly. Remote collection
-// discovery is an account concern; New Local Calendar remains a local-only flow.
-func NewAccountDialogModel(theme Theme) CalendarDialogModel {
-	dialog := NewDialog("Add Account", DefaultDialogStyles())
-	dialog.SetWidth(62)
-	form := newCalDAVConnectionForm(theme, "")
-	m := CalendarDialogModel{
-		dialog:            dialog,
-		form:              form,
-		body:              viewport.New(),
-		help:              newThemedHelp(theme),
-		theme:             theme,
-		accentColor:       theme.Selected,
-		mutedColor:        theme.Muted,
-		textDimColor:      theme.TextDim,
-		accountConnection: true,
-	}
-	m.body.MouseWheelEnabled = true
-	return m
-}
-
-func newCalDAVConnectionForm(theme Theme, usernamePrefill string) Form {
-	styles := DefaultFormStyles()
-	styles.LabelLayout = LabelInline
-	styles.ShowFocusMarker = true
-	styles.ButtonAlign = ButtonAlignRight
-	styles.ButtonRule = true
-
-	insecure := NewCheckboxField("", false)
-	insecure.SetContent("allow plain HTTP")
-	form := NewForm("Sign In", styles,
-		FormItem{Label: "Server URL", Field: newRemoteURLField(""), Required: true},
-		FormItem{Label: "Username", Field: newUsernameField(usernamePrefill), Required: true},
-		FormItem{Label: "Auth", Field: newAuthField("basic"), Required: true},
-		FormItem{Label: "Password", Field: newPasswordField(), Required: true},
-		FormItem{Label: "HTTP", Field: insecure},
-	)
-	form.SetActionButton("Test", Button, func() tea.Msg {
-		return testConnectionPressedMsg{}
-	})
-	form.OnCancel(func(*Form) tea.Cmd {
-		return func() tea.Msg { return CalendarDialogClosedMsg{} }
-	})
-
-	var snapshot struct {
-		secret, clientID, clientSecret string
-		allowInsecure                  bool
-	}
-	oauthLayout := new(bool)
-	snapshotTail := func(f *Form) {
-		if *oauthLayout {
-			snapshot.clientID = f.Field(calDAVIdxOAuthClientID).(*TextField).Value()
-			snapshot.clientSecret = f.Field(calDAVIdxOAuthClientSecret).(*TextField).Value()
-			snapshot.allowInsecure = f.Field(calDAVIdxOAuthAllowInsecure).(*CheckboxField).Checked()
-			return
-		}
-		snapshot.secret = f.Field(calDAVIdxSecret).(*TextField).Value()
-		snapshot.allowInsecure = f.Field(calDAVIdxAllowInsecure).(*CheckboxField).Checked()
-	}
-	appendTail := func(f *Form, authType string) {
-		allow := NewCheckboxField("", snapshot.allowInsecure)
-		allow.SetContent("allow plain HTTP")
-		if calendarAuthIsOAuth(authType) {
-			clientSecret := newOAuthClientSecretField()
-			clientSecret.SetValue(snapshot.clientSecret)
-			f.AppendItems(
-				FormItem{Label: "Client ID", Field: newOAuthClientIDField(snapshot.clientID), Required: true},
-				FormItem{Label: "Client secret", Field: clientSecret, Required: true},
-				FormItem{Label: "HTTP", Field: allow},
-			)
-			*oauthLayout = true
-			return
-		}
-		secret := newPasswordField()
-		secret.SetValue(snapshot.secret)
-		f.AppendItems(
-			FormItem{Label: "Password", Field: secret, Required: true},
-			FormItem{Label: "HTTP", Field: allow},
-		)
-		*oauthLayout = false
-	}
-	form.OnRebuild(func(f *Form) {
-		authType := f.Field(calDAVIdxAuth).(*SelectField).Value()
-		if calendarAuthIsOAuth(authType) != *oauthLayout {
-			snapshotTail(f)
-			f.RemoveItems(calDAVIdxSecret)
-			f.ClearError()
-			appendTail(f, authType)
-		}
-		if !*oauthLayout {
-			secret := f.Field(calDAVIdxSecret).(*TextField)
-			if authType == "bearer" {
-				f.SetItemLabel(calDAVIdxSecret, "Token")
-				secret.SetPlaceholder("paste your API token")
-			} else {
-				f.SetItemLabel(calDAVIdxSecret, "Password")
-				secret.SetPlaceholder("your password")
-			}
-		}
-
-		insecureIdx := calDAVIdxAllowInsecure
-		if *oauthLayout {
-			insecureIdx = calDAVIdxOAuthAllowInsecure
-		}
-		allow := f.Field(insecureIdx).(*CheckboxField)
-		wasAuto := allow.AutoChecked()
-		if isLocalhostHTTP(f.Field(calDAVIdxServer).(*TextField).Value()) {
-			allow.SetChecked(true)
-			allow.SetAutoChecked(true)
-			allow.SetSuffix("")
-			allow.SetDisabledWhen(func() (bool, string) {
-				return true, lipgloss.NewStyle().Foreground(theme.Muted).Italic(true).
-					Render("auto-enabled for localhost")
-			})
-		} else {
-			if wasAuto {
-				allow.SetChecked(false)
-				allow.SetAutoChecked(false)
-			}
-			allow.SetDisabledWhen(nil)
-			if allow.Checked() {
-				allow.SetSuffix(lipgloss.NewStyle().Foreground(theme.Error).Render("(unencrypted)"))
-			} else {
-				allow.SetSuffix("")
-			}
-		}
-	})
-	form.OnSubmit(func(f *Form) tea.Cmd {
-		msg := CalendarDiscoveryRequestedMsg{
-			ServerURL: strings.TrimSpace(f.Field(calDAVIdxServer).(*TextField).Value()),
-			Username:  strings.TrimSpace(f.Field(calDAVIdxUsername).(*TextField).Value()),
-			AuthType:  f.Field(calDAVIdxAuth).(*SelectField).Value(),
-		}
-		if msg.ServerURL == "" {
-			f.SetError(calDAVIdxServer, "Server URL is required")
-			return nil
-		}
-		if msg.Username == "" {
-			f.SetError(calDAVIdxUsername, "Username is required")
-			return nil
-		}
-		if calendarAuthIsOAuth(msg.AuthType) {
-			msg.OAuthClientID = strings.TrimSpace(f.Field(calDAVIdxOAuthClientID).(*TextField).Value())
-			msg.OAuthClientSecret = strings.TrimSpace(f.Field(calDAVIdxOAuthClientSecret).(*TextField).Value())
-			msg.AllowInsecure = f.Field(calDAVIdxOAuthAllowInsecure).(*CheckboxField).Checked()
-			if msg.OAuthClientID == "" {
-				f.SetError(calDAVIdxOAuthClientID, "Client ID is required")
-				return nil
-			}
-			if msg.OAuthClientSecret == "" {
-				f.SetError(calDAVIdxOAuthClientSecret, "Client secret is required")
-				return nil
-			}
-		} else {
-			msg.Secret = f.Field(calDAVIdxSecret).(*TextField).Value()
-			msg.AllowInsecure = f.Field(calDAVIdxAllowInsecure).(*CheckboxField).Checked()
-			if strings.TrimSpace(msg.Secret) == "" {
-				f.SetError(calDAVIdxSecret, "Credential is required")
-				return nil
-			}
-		}
-		return func() tea.Msg { return msg }
-	})
-	return form
 }
 
 // isLocalhostHTTP reports whether a URL uses http:// against localhost
@@ -918,340 +503,5 @@ func (m CalendarDialogModel) leftMovesCursor() bool {
 		return true
 	default:
 		return false
-	}
-}
-
-func (m CalendarDialogModel) SetSize(w, h int) CalendarDialogModel {
-	m.dialog = m.dialog.Update(tea.WindowSizeMsg{Width: w, Height: h})
-	if m.discoveryPicker != nil {
-		picker := m.discoveryPicker.SetSize(w, h)
-		m.discoveryPicker = &picker
-		return m
-	}
-	m.form.SetWidth(m.dialog.ContentWidth())
-	if m.contentWidth != nil {
-		*m.contentWidth = m.dialog.ContentWidth()
-	}
-	m.syncBodyViewport(true)
-	return m
-}
-
-func (m CalendarDialogModel) formViewportHeight() int {
-	const chromeLines = 2 + // top + bottom border
-		1 + // top padding (PaddingY)
-		2 + // dialog title + blank line
-		2 // blank line + help footer
-	extra := 0
-	if m.testStatus != "" {
-		extra = 1
-	}
-	actionLines := 1 + max(lipgloss.Height(m.form.ButtonRowView()), 1) // separator + buttons
-	return max(m.dialog.height-chromeLines-actionLines-extra, 1)
-}
-
-func (m *CalendarDialogModel) syncBodyViewport(keepFocusVisible bool) {
-	cw := m.dialog.ContentWidth()
-	if cw <= 0 || m.dialog.height <= 0 {
-		return
-	}
-	bodyLines := strings.Split(m.form.BodyView(), "\n")
-	m.body.SetWidth(cw)
-	m.body.SetHeight(min(len(bodyLines), m.formViewportHeight()))
-	m.body.SetContentLines(bodyLines)
-	if keepFocusVisible {
-		m.keepFocusedFieldVisible()
-	}
-}
-
-func (m *CalendarDialogModel) keepFocusedFieldVisible() {
-	if m.body.Height() <= 0 {
-		return
-	}
-	line := m.form.FocusedLine()
-	if line < 0 {
-		// Focus is on the button row, not a body field; leave the
-		// scroll position where the last field left it.
-		return
-	}
-	if line < m.body.YOffset() {
-		m.body.ScrollUp(m.body.YOffset() - line)
-		return
-	}
-	bottom := m.body.YOffset() + m.body.Height() - 1
-	if line > bottom {
-		m.body.ScrollDown(line - bottom)
-	}
-}
-
-// SetInspectorSize prepares the stored form for a borderless render inside
-// the Calendars manager. Render stays pure. Body content and viewport
-// dimensions are refreshed here and after Update, never from InspectorView.
-// Blur returns a copy whose form holds no keyboard focus. The manager can
-// then render it as the root selection preview while the source list owns
-// input.
-func (m CalendarDialogModel) Blur() CalendarDialogModel {
-	m.form = m.form.Blur()
-	return m
-}
-
-func (m CalendarDialogModel) SetInspectorSize(w, h int) CalendarDialogModel {
-	w = max(w, 1)
-	h = max(h, 1)
-	m.form.SetWidth(w)
-	if m.contentWidth != nil {
-		*m.contentWidth = w
-	}
-	bodyLines := strings.Split(m.form.BodyView(), "\n")
-	statusLines := 0
-	if m.testStatus != "" {
-		statusLines = 1
-	}
-	buttonLines := max(lipgloss.Height(m.form.ButtonRowView()), 1)
-	bodyHeight := max(h-2-statusLines-1-buttonLines, 1)
-	m.body.SetWidth(w)
-	m.body.SetHeight(min(len(bodyLines), bodyHeight))
-	m.body.SetContentLines(bodyLines)
-	m.keepFocusedFieldVisible()
-	if m.discoveryPicker != nil {
-		picker := m.discoveryPicker.SetSize(w, h)
-		m.discoveryPicker = &picker
-	}
-	return m
-}
-
-// InspectorView renders the calendar/add-account form without another dialog
-// border so the manager's grouped hierarchy remains mounted beside it.
-func (m CalendarDialogModel) InspectorView(w, h int) string {
-	if m.discoveryPicker != nil {
-		return padLines(strings.Split(m.discoveryPicker.View(), "\n"), w, h)
-	}
-	parts := []string{lipgloss.NewStyle().Bold(true).Render(truncateTo(m.dialog.title, w)), m.body.View()}
-	if m.testStatus != "" {
-		parts = append(parts, truncateTo(m.testStatus, w))
-	}
-	parts = append(parts, m.actionsSeparator(w), m.form.ButtonRowView())
-	return padLines(strings.Split(strings.Join(parts, "\n"), "\n"), w, h)
-}
-
-// HelpBindings returns the footer help this pushed detail owns while it is
-// the manager's active screen. An embedded discovery picker advertises its
-// own compact set, otherwise the form's field keys. The manager footer
-// defers here so a rebound key no longer drifts between the child and the
-// footer. These are display-only. Input route is unchanged.
-func (m CalendarDialogModel) HelpBindings() []key.Binding {
-	if m.discoveryPicker != nil {
-		return m.discoveryPicker.HelpBindings()
-	}
-	return []key.Binding{
-		footerBinding("tab", "next field"),
-		footerBinding("enter", "confirm"),
-		footerBinding("esc", "back"),
-	}
-}
-
-func (m CalendarDialogModel) bodyOverflows() bool {
-	return m.body.TotalLineCount() > m.body.VisibleLineCount()
-}
-
-func (m CalendarDialogModel) scrollHint() string {
-	if !m.bodyOverflows() {
-		return ""
-	}
-	switch {
-	case m.body.AtTop():
-		return "↓ more"
-	case m.body.AtBottom():
-		return "↑ more"
-	default:
-		return "↑↓ more"
-	}
-}
-
-func (m CalendarDialogModel) actionsSeparator(w int) string {
-	faint := lipgloss.NewStyle().Faint(true)
-	hint := m.scrollHint()
-	hw := lipgloss.Width(hint)
-	if hint == "" || w <= hw+2 {
-		return faint.Render(strings.Repeat("─", w))
-	}
-	left := (w - hw - 2) / 2
-	right := w - hw - 2 - left
-	return faint.Render(strings.Repeat("─", left)) + " " + faint.Render(hint) + " " + faint.Render(strings.Repeat("─", right))
-}
-
-func (m CalendarDialogModel) BoxSize() (int, int) {
-	return lipgloss.Size(m.View())
-}
-
-func (m CalendarDialogModel) Update(msg tea.Msg) (CalendarDialogModel, tea.Cmd) {
-	if msg, ok := msg.(tea.WindowSizeMsg); ok {
-		return m.SetSize(msg.Width, msg.Height), nil
-	}
-
-	if m.discoveryPicker != nil {
-		picker, cmd := m.discoveryPicker.Update(msg)
-		m.discoveryPicker = &picker
-		return m, cmd
-	}
-
-	if _, ok := msg.(testConnectionPressedMsg); ok {
-		m, cmd := m.handleTestPressed()
-		m.syncBodyViewport(true)
-		return m, cmd
-	}
-
-	if _, ok := msg.(calendarSavePromotePressedMsg); ok {
-		if m.saveMakeDefault != nil {
-			*m.saveMakeDefault = true
-		}
-		var cmd tea.Cmd
-		m.form, cmd = m.form.Submit()
-		return m, cmd
-	}
-
-	if tr, ok := msg.(CalendarTestResultMsg); ok {
-		if tr.OK {
-			m.testStatus = lipgloss.NewStyle().Foreground(m.theme.Accent).
-				Render("✓ " + tr.Message)
-		} else {
-			m.testStatus = lipgloss.NewStyle().Foreground(m.theme.Error).
-				Render("✗ " + tr.Message)
-		}
-		m.syncBodyViewport(true)
-		return m, nil
-	}
-
-	if msg, ok := msg.(tea.KeyPressMsg); ok {
-		switch {
-		case key.Matches(msg, key.NewBinding(key.WithKeys("esc"))):
-			return m, func() tea.Msg { return CalendarDialogClosedMsg{} }
-		case key.Matches(msg, key.NewBinding(key.WithKeys("ctrl+s"))):
-			var cmd tea.Cmd
-			m.form, cmd = m.form.Submit()
-			return m, cmd
-		}
-	}
-
-	if mc, ok := msg.(tea.MouseClickMsg); ok {
-		if mc.Button == tea.MouseLeft {
-			bw, bh := m.BoxSize()
-			ox := (m.dialog.width - bw) / 2
-			oy := (m.dialog.height - bh) / 2
-			target := mouseResolve(mc.X-ox, mc.Y-oy)
-			// A click on the Display Calendar checkbox toggles it inside the
-			// form. Compare pre/post state. The mouse path then emits the
-			// same CalendarVisibilityToggledMsg as the keyboard path.
-			preVisible := m.visibilityChecked()
-			var cmd tea.Cmd
-			m.form, cmd = m.form.Update(MouseEvent{IsClick: true, Target: target})
-			m.syncBodyViewport(true)
-			return m.applyVisibilityToggle(preVisible, cmd)
-		}
-		return m, nil
-	}
-	if mw, ok := msg.(tea.MouseWheelMsg); ok {
-		var cmd tea.Cmd
-		m.syncBodyViewport(false)
-		m.body, cmd = m.body.Update(mw)
-		return m, cmd
-	}
-
-	preVisible := m.visibilityChecked()
-	var cmd tea.Cmd
-	m.form, cmd = m.form.Update(msg)
-	m.syncBodyViewport(true)
-	return m.applyVisibilityToggle(preVisible, cmd)
-}
-
-// visibilityChecked reports the Display Calendar checkbox's current state, or
-// true when there is no checkbox (create mode). A no-op comparison then never
-// reports a spurious change.
-func (m CalendarDialogModel) visibilityChecked() bool {
-	if m.visibilityCb == nil {
-		return true
-	}
-	return m.visibilityCb.Checked()
-}
-
-// applyVisibilityToggle compares the checkbox state before and after a form
-// update. A change emits CalendarVisibilityToggledMsg with the DESIRED hidden
-// state. It mirrors it into the local model so the dot flips with no reload.
-// It batches with any cmd the form update produced. Metadata Save/Cancel
-// never persists visibility.
-func (m CalendarDialogModel) applyVisibilityToggle(preVisible bool, cmd tea.Cmd) (CalendarDialogModel, tea.Cmd) {
-	if m.visibilityCb == nil {
-		return m, cmd
-	}
-	postVisible := m.visibilityCb.Checked()
-	if postVisible == preVisible {
-		return m, cmd
-	}
-	m.hidden = !postVisible
-	id := m.id
-	toggle := func() tea.Msg { return CalendarVisibilityToggledMsg{ID: id, Hidden: !postVisible} }
-	if cmd == nil {
-		return m, toggle
-	}
-	return m, tea.Batch(cmd, toggle)
-}
-
-func (m CalendarDialogModel) View() string {
-	if m.discoveryPicker != nil {
-		return m.discoveryPicker.View()
-	}
-	helpKeys := []key.Binding{
-		key.NewBinding(key.WithKeys("tab"), key.WithHelp("tab", "next field")),
-		key.NewBinding(key.WithKeys("enter"), key.WithHelp("enter", "confirm")),
-		key.NewBinding(key.WithKeys("esc"), key.WithHelp("esc", "cancel")),
-	}
-	m.dialog.SetFooter(m.help.ShortHelpView(helpKeys))
-	m.syncBodyViewport(false)
-	cw := m.dialog.ContentWidth()
-	parts := []string{m.body.View()}
-	if m.testStatus != "" {
-		parts = append(parts, truncateTo(m.testStatus, cw))
-	}
-	parts = append(parts, m.actionsSeparator(cw), m.form.ButtonRowView())
-	body := strings.Join(parts, "\n")
-	content := mouseSweep(m.dialog.Box(body))
-	return content
-}
-
-// handleTestPressed validates the remote fields. When they have values,
-// it emits a CalendarTestRequestedMsg so the parent can run the
-// authenticated ping. Errors show inline. The server is not contacted.
-func (m CalendarDialogModel) handleTestPressed() (CalendarDialogModel, tea.Cmd) {
-	if !m.accountConnection {
-		return m, nil
-	}
-	// The oauth2 layout has no password to ping with — there is no token
-	// until the browser flow runs, which happens on sign-in.
-	if calendarAuthIsOAuth(m.form.Field(calDAVIdxAuth).(*SelectField).Value()) {
-		m.testStatus = lipgloss.NewStyle().Foreground(m.theme.TextDim).Italic(true).
-			Render("Connection test runs after Google authorization — sign in to continue")
-		return m, nil
-	}
-	url := strings.TrimSpace(m.form.Field(calDAVIdxServer).(*TextField).Value())
-	user := strings.TrimSpace(m.form.Field(calDAVIdxUsername).(*TextField).Value())
-	auth := m.form.Field(calDAVIdxAuth).(*SelectField).Value()
-	pass := m.form.Field(calDAVIdxSecret).(*TextField).Value()
-	ins := m.form.Field(calDAVIdxAllowInsecure).(*CheckboxField).Checked()
-
-	if url == "" || user == "" || pass == "" {
-		m.testStatus = lipgloss.NewStyle().Foreground(m.theme.Error).
-			Render("✗ Fill URL, Username, and Password first")
-		return m, nil
-	}
-
-	m.testStatus = lipgloss.NewStyle().Foreground(m.theme.TextDim).Italic(true).
-		Render("Testing…")
-	return m, func() tea.Msg {
-		return CalendarTestRequestedMsg{
-			URL:           url,
-			Username:      user,
-			AuthType:      auth,
-			Password:      pass,
-			AllowInsecure: ins,
-		}
 	}
 }
